+++ src/Bus/Middleware/SetupBeforeDispatch.php
@@ -20,7 +20,7 @@ declare(strict_types=1);
 namespace CloudCreativity\Modules\Bus\Middleware;
 
 use Closure;
-use CloudCreativity\Modules\Bus\MessageInterface;
+use CloudCreativity\Modules\Toolkit\Messages\MessageInterface;
 use CloudCreativity\Modules\Toolkit\Result\ResultInterface;
 
 final class SetupBeforeDispatch implements MessageMiddlewareInterface